+++ src/pages/Main.tsx
@@ -1,4 +1,3 @@
-// import { Link } from "react-router-dom";
 import Header from "../components/Header";
 import Home from "../components/Home";
